removed extensions folder and namespace, fixed product bundle properties
